+++ src/test/java/view/ContinuePlayTest.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
